Implement sending logic
Implemented the sending logic to package the data into Json objects and then send them. The Client receives the initial state, then perpetually updates the server with its interpretation of the game state. If the server sees a discrepancy with the sync IDs between both players, and the player being examined is behind, then it gets its state overwritten and sent back. If not, then the sent data is parroted back to the client.